+++ Assets/Prefabs/Coglings/Buildings/Cogling_Gate.prefab
@@ -12,6 +12,8 @@ GameObject:
   - component: {fileID: 5688697115624436389}
   - component: {fileID: 7645544545388728891}
   - component: {fileID: 7305483904886340058}
+  - component: {fileID: 6525835175647597740}
+  - component: {fileID: 1236551294335420954}
   m_Layer: 0
   m_Name: Cogling_Gate
   m_TagString: EnemyBuilding
@@ -88,7 +90,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 6706871883591336528}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &6525835175647597740
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2668744761997871384}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &1236551294335420954
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2668744761997871384}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &3254225055397663186
 GameObject:
   m_ObjectHideFlags: 0
