+++ Assets/Prefabs/Coglings/Buildings/Cogling_House_1.prefab
@@ -4669,6 +4669,8 @@ GameObject:
   - component: {fileID: 8411656157473492728}
   - component: {fileID: 5670007089494075935}
   - component: {fileID: 1850607732582464769}
+  - component: {fileID: 3541655678656809498}
+  - component: {fileID: 583252692135844577}
   m_Layer: 0
   m_Name: Cogling_House_1
   m_TagString: EnemyBuilding
@@ -4757,7 +4759,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 1673920813656395620}
   currentHealth: 0
-  maxHealth: 500
+  maxHealth: 0
+--- !u!114 &3541655678656809498
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1942300920393013544}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &583252692135844577
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1942300920393013544}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &4483718374873311356
 GameObject:
   m_ObjectHideFlags: 0
