+++ Assets/Prefabs/Coglings/Buildings/Cogling_Resource_Depot.prefab
@@ -182,6 +182,8 @@ GameObject:
   - component: {fileID: 8440552994888088031}
   - component: {fileID: 6130455730939591851}
   - component: {fileID: 153134493302275331}
+  - component: {fileID: 2593021657931484584}
+  - component: {fileID: 2974335991512074337}
   m_Layer: 0
   m_Name: Cogling_Resource_Depot
   m_TagString: EnemyBuilding
@@ -285,7 +287,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 6788983062589977070}
   currentHealth: 0
-  maxHealth: 500
+  maxHealth: 0
+--- !u!114 &2593021657931484584
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7362237247213061779}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &2974335991512074337
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7362237247213061779}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &7611906852141158259
 GameObject:
   m_ObjectHideFlags: 0
