+++ Assets/Prefabs/Coglings/Buildings/Cogling_Town_Center.prefab
@@ -743,6 +743,8 @@ GameObject:
   - component: {fileID: 5957536318416629450}
   - component: {fileID: 1534940599697223844}
   - component: {fileID: 5920074685392764735}
+  - component: {fileID: 8173644313400047849}
+  - component: {fileID: 2515512827970197295}
   m_Layer: 0
   m_Name: Cogling_Town_Center
   m_TagString: EnemyBuilding
@@ -830,7 +832,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 614439471125923004}
   currentHealth: 0
-  maxHealth: 2000
+  maxHealth: 0
+--- !u!114 &8173644313400047849
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 78809170135114686}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &2515512827970197295
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 78809170135114686}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 8
 --- !u!1 &78809170375100892
 GameObject:
   m_ObjectHideFlags: 0
