+++ Assets/Prefabs/Coglings/Buildings/Cogling_Wall_No_Column.prefab
@@ -12,6 +12,8 @@ GameObject:
   - component: {fileID: 8462652319728074355}
   - component: {fileID: 789307896517772038}
   - component: {fileID: 4609213276799584311}
+  - component: {fileID: 4935132064023739305}
+  - component: {fileID: 6093382977574822805}
   m_Layer: 0
   m_Name: Cogling_Wall_No_Column
   m_TagString: EnemyBuilding
@@ -83,7 +85,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 1352282471178123724}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &4935132064023739305
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 520668927584109619}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &6093382977574822805
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 520668927584109619}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &691644960548480496
 GameObject:
   m_ObjectHideFlags: 0
