+++ Assets/Prefabs/Graxxians/Buildings/Building Parts/Graxxian_Wall.prefab
@@ -163,6 +163,8 @@ GameObject:
   - component: {fileID: 8746219490940178552}
   - component: {fileID: 1832408016891661561}
   - component: {fileID: 391130005409506867}
+  - component: {fileID: 976789841129243053}
+  - component: {fileID: 8375573968459105292}
   m_Layer: 0
   m_Name: Graxxian_Wall
   m_TagString: EnemyBuilding
@@ -234,7 +236,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 1338836299113706709}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &976789841129243053
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4400036841116834414}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &8375573968459105292
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4400036841116834414}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &7898037224852863272
 GameObject:
   m_ObjectHideFlags: 0
