+++ Assets/Prefabs/Graxxians/Buildings/Building Parts/Graxxian_Wall_No_Edge.prefab
@@ -12,6 +12,8 @@ GameObject:
   - component: {fileID: 858601406834335983}
   - component: {fileID: 6070307586862839451}
   - component: {fileID: 1847987234133189960}
+  - component: {fileID: 1963084502758888117}
+  - component: {fileID: 2099251230477434836}
   m_Layer: 0
   m_Name: Graxxian_Wall_No_Edge
   m_TagString: EnemyBuilding
@@ -82,7 +84,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 6017530036704197441}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &1963084502758888117
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2788020826972690051}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &2099251230477434836
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2788020826972690051}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &3760211653283961988
 GameObject:
   m_ObjectHideFlags: 0
