+++ Assets/Prefabs/Graxxians/Buildings/Graxxian_House.prefab
@@ -202,6 +202,8 @@ GameObject:
   - component: {fileID: 1024673932325785468}
   - component: {fileID: 6621353052968863107}
   - component: {fileID: 3027916033893768809}
+  - component: {fileID: -6088600252783369005}
+  - component: {fileID: 4894939263274480973}
   m_Layer: 0
   m_Name: Graxxian_House
   m_TagString: EnemyBuilding
@@ -298,7 +300,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 5301502733302480193}
   currentHealth: 0
-  maxHealth: 500
+  maxHealth: 0
+--- !u!114 &-6088600252783369005
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3570326620244428714}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &4894939263274480973
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3570326620244428714}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &7061513270263491555
 GameObject:
   m_ObjectHideFlags: 0
