+++ Assets/Prefabs/Robot Empire/Buildings/Base_Defenses_Building_Template.prefab
@@ -12,6 +12,8 @@ GameObject:
   - component: {fileID: 1466432658344948982}
   - component: {fileID: 6315394591933485129}
   - component: {fileID: 184252387293389671}
+  - component: {fileID: 4319256136116311559}
+  - component: {fileID: 1999632849276059101}
   m_Layer: 0
   m_Name: Base_Defenses_Building_Template
   m_TagString: Untagged
@@ -79,7 +81,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 2718181009849247439}
   currentHealth: 0
-  maxHealth: 1000
+  maxHealth: 0
+--- !u!114 &4319256136116311559
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1987490933575897882}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
+--- !u!114 &1999632849276059101
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1987490933575897882}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
 --- !u!1 &2813859042357552073
 GameObject:
   m_ObjectHideFlags: 0
