+++ Assets/Prefabs/Robot Empire/Buildings/Troop_Cap_Building_Template.prefab
@@ -298,6 +298,8 @@ GameObject:
   - component: {fileID: 4477807023332532627}
   - component: {fileID: -5974618585579986154}
   - component: {fileID: 3008651519042365280}
+  - component: {fileID: 1204881041888420525}
+  - component: {fileID: 782814296953686794}
   m_Layer: 0
   m_Name: Troop_Cap_Building_Template
   m_TagString: Untagged
@@ -352,7 +354,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 6208785234820759101}
   currentHealth: 0
-  maxHealth: 1000
+  maxHealth: 0
+--- !u!114 &1204881041888420525
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5067692701915907592}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &782814296953686794
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5067692701915907592}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &7507308112042898650
 GameObject:
   m_ObjectHideFlags: 0
