+++ Assets/Prefabs/Robot Empire/Buildings/Troop_Production_Building_Template.prefab
@@ -262,6 +262,8 @@ GameObject:
   - component: {fileID: 6187068616382062810}
   - component: {fileID: 2367912232560805644}
   - component: {fileID: 4991531160258112648}
+  - component: {fileID: 8700735092147369950}
+  - component: {fileID: 8775919825527588784}
   m_Layer: 0
   m_Name: Troop_Production_Building_Template
   m_TagString: Untagged
@@ -316,7 +318,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 2173473180147627375}
   currentHealth: 0
-  maxHealth: 1000
+  maxHealth: 0
+--- !u!114 &8700735092147369950
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6187068616382062813}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &8775919825527588784
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6187068616382062813}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &6187068616687675348
 GameObject:
   m_ObjectHideFlags: 0
