+++ Assets/Prefabs/Sleemasi/Buildings/Building Parts/Sleemasi_Wall.prefab
@@ -12,6 +12,8 @@ GameObject:
   - component: {fileID: 5918334636125100728}
   - component: {fileID: 7350685825568788550}
   - component: {fileID: 6640459710909466622}
+  - component: {fileID: 2440114558854783882}
+  - component: {fileID: 7965466810253888569}
   m_Layer: 0
   m_Name: Sleemasi_Wall
   m_TagString: EnemyBuilding
@@ -79,7 +81,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 4161990635377430384}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &2440114558854783882
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1363302006465784908}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &7965466810253888569
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1363302006465784908}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &2817239426916843374
 GameObject:
   m_ObjectHideFlags: 0
