+++ Assets/Prefabs/Sleemasi/Buildings/Building Parts/Sleemasi_Wall_No_Corner.prefab
@@ -87,6 +87,8 @@ GameObject:
   - component: {fileID: 7523835118532942311}
   - component: {fileID: 7162548314340996398}
   - component: {fileID: 328409946981275334}
+  - component: {fileID: 8232480288669954291}
+  - component: {fileID: 1663328706204657446}
   m_Layer: 0
   m_Name: Sleemasi_Wall_No_Corner
   m_TagString: EnemyBuilding
@@ -153,7 +155,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 4442439873701893593}
   currentHealth: 0
-  maxHealth: 1500
+  maxHealth: 0
+--- !u!114 &8232480288669954291
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3658491172958650253}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &1663328706204657446
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3658491172958650253}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 11
 --- !u!1 &4380018040761950434
 GameObject:
   m_ObjectHideFlags: 0
