+++ Assets/Prefabs/Sleemasi/Buildings/Sleemasi_House_3.prefab
@@ -88,6 +88,8 @@ GameObject:
   - component: {fileID: 6254126654253459008}
   - component: {fileID: 4885819153746341337}
   - component: {fileID: 5461540792715532847}
+  - component: {fileID: 6178793410200540970}
+  - component: {fileID: 4839113914717035623}
   m_Layer: 0
   m_Name: Sleemasi_House_3
   m_TagString: EnemyBuilding
@@ -164,7 +166,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 1404433941004856630}
   currentHealth: 0
-  maxHealth: 500
+  maxHealth: 0
+--- !u!114 &6178793410200540970
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4117881052247255980}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &4839113914717035623
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4117881052247255980}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 10
 --- !u!1 &4420681350814083400
 GameObject:
   m_ObjectHideFlags: 0
