+++ Assets/Prefabs/Sleemasi/Buildings/Sleemasi_Town_Center.prefab
@@ -2742,6 +2742,8 @@ GameObject:
   - component: {fileID: 3165440000840267492}
   - component: {fileID: 3597080273506042354}
   - component: {fileID: 3441697372896030449}
+  - component: {fileID: 4475766793936918130}
+  - component: {fileID: 4956357769855406632}
   m_Layer: 0
   m_Name: Sleemasi_Town_Center
   m_TagString: EnemyBuilding
@@ -2902,7 +2904,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 4330509824106338987}
   currentHealth: 0
-  maxHealth: 2000
+  maxHealth: 0
+--- !u!114 &4475766793936918130
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2757575571955915716}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &4956357769855406632
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2757575571955915716}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 8
 --- !u!1 &2984978330860748247
 GameObject:
   m_ObjectHideFlags: 0
