+++ Assets/Prefabs/Sleemasi/Buildings/sleemasi_turret.prefab
@@ -182,6 +182,8 @@ GameObject:
   - component: {fileID: 7994315598192679485}
   - component: {fileID: -6835098724184767142}
   - component: {fileID: 4849170518120719512}
+  - component: {fileID: 3920952309511562758}
+  - component: {fileID: 2940689782901007275}
   m_Layer: 0
   m_Name: sleemasi_turret
   m_TagString: EnemyBuilding
@@ -365,7 +367,33 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   healthbar: {fileID: 6524798561653927208}
   currentHealth: 0
-  maxHealth: 1250
+  maxHealth: 0
+--- !u!114 &3920952309511562758
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3407446719780975775}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2907ce02e59e5f0439b22e2497fa465e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Prefab: {fileID: 0}
+--- !u!114 &2940689782901007275
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3407446719780975775}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88ec9ecfb8a556d4684d045ebea34837, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  unit_Name: 9
 --- !u!1 &3728675082817132538
 GameObject:
   m_ObjectHideFlags: 0
