+++ Assets/Scripts/Buildings/House_Stats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 678c2d3014d82c140a7e4d7b96295cf4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
