+++ Assets/Scripts/Buildings/Turret_Stats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 590dccf570f22124293a462cefc799eb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
