+++ Assets/Scripts/Buildings/WallGate_Stats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2e2438e719426344a759c343c466522
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
