--- Assets/Scripts/Main_Base.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a3622525abe2394c80dca3b16430690
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
