--- Assets/Scripts/Main_Base/Main_Base_Stats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b48a28da2721ea948acd5340de0d6b29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
